refactor: grouping jsonld components
