+++ components/json-ld/website.tsx
@@ -1,10 +1,10 @@
-import { JsonLdBuilder } from "./jsonld-builder";
+import { JsonLdBuilder } from "./builder";
 import {
   makeOrganizationJsonLd,
   makeWebsiteJsonLd,
 } from "~/lib/jsonld-generator";
 
-export function WebsiteJsonLd() {
+export function JsonLdWebsite() {
   return JsonLdBuilder({
     jsonsLd: [makeOrganizationJsonLd(), makeWebsiteJsonLd()],
   });
